+++ core/src/commonMain/kotlin/com/saveourtool/processbuilder/utils/PlatformUtils.kt
@@ -4,6 +4,7 @@
 
 @file:Suppress("FILE_NAME_MATCH_CLASS", "MatchingDeclarationName")
 @file:JvmName("PlatformUtilsJVM")
+
 package com.saveourtool.processbuilder.utils
 
 import kotlin.jvm.JvmName
