+++ core/src/commonNonJsTest/kotlin/com/saveourtool/processbuilder/ProcessBuilderTest.kt
@@ -2,9 +2,7 @@ package com.saveourtool.processbuilder
 
 import com.saveourtool.processbuilder.ProcessBuilder.Companion.processCommandWithEcho
 import com.saveourtool.processbuilder.exceptions.ProcessExecutionException
-import com.saveourtool.processbuilder.utils.CurrentOs
 import com.saveourtool.processbuilder.utils.fs
-import com.saveourtool.processbuilder.utils.getCurrentOs
 import com.saveourtool.processbuilder.utils.isCurrentOsWindows
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -120,35 +118,5 @@ class ProcessBuilderTest {
         val inputCommand = "echo foo bar ; echo b; ls"
         val expectedCommand = "echo | set /p dummyName=\"foo bar\"  ; echo | set /p dummyName=\"b\"  ; ls"
         assertEquals(expectedCommand, processCommandWithEcho(inputCommand))
-    }
-
-    @Test
-    fun `check stderr`() {
-        val actualResult = processBuilder.exec("cd non_existent_dir", "", null, 10_000L)
-        val expectedStdout: List<String> = emptyList()
-        val (expectedCode, expectedStderr) = when (getCurrentOs()) {
-            CurrentOs.LINUX -> 2 to listOf("sh: 1: cd: can't cd to non_existent_dir")
-            CurrentOs.MACOS -> 1 to listOf("sh: line 0: cd: non_existent_dir: No such file or directory")
-            CurrentOs.WINDOWS -> 1 to listOf("The system cannot find the path specified.")
-            else -> return
-        }
-        assertEquals(expectedCode, actualResult.code)
-        assertEquals(expectedStdout, actualResult.stdout)
-        assertEquals(expectedStderr, actualResult.stderr)
-    }
-
-    @Test
-    fun `check stderr with additional warning`() {
-        val actualResult = processBuilder.exec("cd non_existent_dir 2>/dev/null", "", null, 10_000L)
-        val expectedStdout: List<String> = emptyList()
-        val (expectedCode, expectedStderr) = when (getCurrentOs()) {
-            CurrentOs.LINUX -> 2 to emptyList()
-            CurrentOs.MACOS -> 1 to emptyList()
-            CurrentOs.WINDOWS -> 1 to listOf("The system cannot find the path specified.")
-            else -> return
-        }
-        assertEquals(expectedCode, actualResult.code)
-        assertEquals(expectedStdout, actualResult.stdout)
-        assertEquals(expectedStderr, actualResult.stderr)
     }
 }
